Move MERGE resolution from catalyst to core
