+++ spark/src/main/scala/org/apache/spark/sql/delta/DeltaAnalysis.scala
@@ -535,7 +535,7 @@ class DeltaAnalysis(session: SparkSession)
           matchedActions ++ notMatchedActions ++ notMatchedBySourceActions
         )
 
-        DeltaMergeInto.resolveReferencesAndSchema(deltaMerge, conf)(
+        ResolveDeltaMergeInto.resolveReferencesAndSchema(deltaMerge, conf)(
           tryResolveReferencesForExpressions(session))
       } else {
         merge
@@ -577,7 +577,7 @@ class DeltaAnalysis(session: SparkSession)
 
     case deltaMerge: DeltaMergeInto =>
       val d = if (deltaMerge.childrenResolved && !deltaMerge.resolved) {
-        DeltaMergeInto.resolveReferencesAndSchema(deltaMerge, conf)(
+        ResolveDeltaMergeInto.resolveReferencesAndSchema(deltaMerge, conf)(
           tryResolveReferencesForExpressions(session))
       } else deltaMerge
       d.copy(target = stripTempViewForMergeWrapper(d.target))
